Changed top level argument design from providing grid / image size limits to providing grid complexity and aspect.
Changed characters to produce grids with the same aspect as the
original character by default.

Greatly simplified implementation of SuperShapes so that only the inherited
SuperShape needs to be created. Component shapes are created dynamically
based on the data stored in SuperShape.

With explicit sizes, different shape grids yield wildly different graph
dimensions and even more importantly wildly different complexity depending
on the number of edges per shape.

With aspect, all shapes yield a graph with the same shape.
With complexity, grids produce more shapes for simple shapes and fewer
shapes for complex shapes so that the complexity is equivalent.